requirements.txt
py-canvas-api>=0.2.7
